requirements_ui.txt
gradio
pdfplumber
sentence-transformers
genanki
groq
google-genai
transformers
torch